+++ src/Exceptions/CheckoutFinalizationException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Udviklr\CashierNets\Exceptions;
+
+use RuntimeException;
+
+final class CheckoutFinalizationException extends RuntimeException
+{
+}